api/network/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
ipaddress==1.0.23
pytest==7.4.3
gunicorn==21.2.0